server/server/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cors==1.0.1
filelock==3.25.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-marshmallow==1.4.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.9.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
PySocks==1.7.1
python-dotenv==1.2.2
python-slugify==8.0.4
redis==7.3.0
requests==2.32.5
requests-file==3.0.1
SQLAlchemy==2.0.48
text-unidecode==1.3
tldextract==5.3.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zope.event==6.1
zope.interface==8.2
